10216-implemented review comments and few home screen cosmetic changes AB#10216 AB#10201
