+++ README.md
@@ -20,7 +20,7 @@ Run `ng build` to build the project. The build artifacts will be stored in the `
 
 ## Running unit tests
 
-Run `npm test` to execute the unit tests via [Jest].
+Run `npm test` to execute the unit tests via [Jest](https://jestjs.io/).
 
 ## Further help
