+++ UnitTests/app.component.spec.ts
@@ -1,0 +1,10 @@
+import { AppComponent } from '../src/app/app.component';
+
+describe('AppComponent', () => {
+  let component: AppComponent;
+
+  it('should exist', () => {
+    component = new AppComponent();
+    expect(component).toBeDefined();
+  })
+})
