+++ constants.ts
@@ -1,0 +1,6 @@
+
+export const constants = {
+    copyright: "@ Crown copyright 2021 UK Hydrographic office",
+    fssTitle: "File Share Service",
+    phase: "alpha"
+};
