+++ package.json
@@ -9,19 +9,25 @@
     "lint": "ng lint",
     "preinstall": "npx npm-force-resolutions"
   },
- "jest-junit": {
+  "jest-junit": {
     "suiteNameTemplate": "{filepath}",
     "outputDirectory": ".",
     "outputName": "junit.xml"
   },
-  "jest":{
+  "jest": {
     "collectCoverage": true,
-     "coverageReporters" :["lcov","html"]
+    "coverageReporters": [
+      "lcov",
+      "html"
+    ]
   },
   "resolutions": {
     "dns-packet": "~5.2.2",
     "postcss": "~8.2.10",
-    "ws": "~7.4.6"
+    "ws": "~7.4.6",
+    "glob-parent": "~5.1.2",
+    "chokidar": "~3.5.1",
+    "css-what": "^5.0.1"
   },
   "private": true,
   "dependencies": {
@@ -36,6 +42,7 @@
     "@angular/router": "~11.2.13",
     "@fortawesome/fontawesome-free": "^5.15.3",
     "@ukho/design-system": "^2.0.0",
+    "jest-junit": "^12.1.0",
     "npm-force-resolutions": "0.0.10",
     "postcss": "^8.3.0",
     "rxjs": "~6.6.0",
@@ -49,8 +56,8 @@
     "@types/jest": "^26.0.23",
     "@types/node": "^12.11.1",
     "codelyzer": "^6.0.0",
-    "jest": "^27.0.3",
-    "jest-preset-angular": "^9.0.1",
+    "jest": "^27.0.4",
+    "jest-preset-angular": "^9.0.3",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
     "typescript": "~4.1.5"
