--- src/app/app.component.spec.ts
@@ -1,10 +1,0 @@
- import { AppComponent } from './app.component';
-
- describe('AppComponent', () => {
-   let component: AppComponent;
-
-   it('should exist', () => {
-     component = new AppComponent();
-     expect(component).toBeDefined();
-   })
- })
