+++ src/app/features/fss-home/fss-home.component.html
@@ -3,7 +3,7 @@
 <ukho-card>
         <p>The ADMIRALTY File Sharing Service allows you to search and download files.</p>
         <h1>Sign in to access your files</h1>
-        <p class="light">You need to sign in with your account to access the File Share Service.</p> 
+        <p><strong class="light">You need to sign in with your account to access the File Share Service.</strong></p>
         <ukho-button>Sign in</ukho-button>
         <p><a href="#" id="fss-createaccount">Create a new account</a></p> 
 </ukho-card>
