+++ src/app/shared/components/fss-footer/fss-footer.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
-import { FooterComponent } from '@ukho/design-system'
-
+import { FooterComponent } from '@ukho/design-system';
+import{ constants } from '../../../../../constants';
 
 @Component({
   selector: 'app-fss-footer',
@@ -17,7 +17,7 @@ export class FssFooterComponent extends FooterComponent {
     this.imageLink="https://www.admiralty.co.uk/";
     this.imageSrc="https://design.ukho.dev/svg/UKHO stacked logo.svg";
     this.imageAlt="Admirality Maritime Data Solutions | UK Hydrographic office";
-    this.text="@ Crown copyright 2021 UK Hydrographic office";
+    this.text=constants.copyright;
     this.navigation = [
       {
         title:"Privacy policy",
