+++ src/styles.scss
@@ -28,7 +28,7 @@
           width: 100%
       }
       p{
-        font-size: 20px;
+        font-size: 25px;
         font-weight: 300;
       }
       a#fss-createaccount{
